edunarrator-backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
PyPDF2
python-docx
google-generativeai
murf-ai 
python-pptx
pdfplumber
